Optimization updates, heat refactor
Optimized the drawing of the map to some degree; more to come (need to probably refactor the whole thing, but most importantly: only redraw a tile when it is necessary to redraw that tile, ie when something has changed)... and added the images. Also refactored the weapon status. Lastly, reset the map when deactivating the hud, so that moving maps doesn't mess up anything.